=General: *Merged missing changes from 3ceam rev. 594 - 595
*Adjusted the homunculus and mvplog tables to support increased EXP amounts.
-Don't forget to run the SQL-Updates.

=Skills:
*RK_DRAGONBREATH damage is now affected by the learned level of RK_DRAGONTRAINING

*NC_PILEBUNKER ranged changed to 1.

*SC_AUTOSHADOWSPELL now autocasts the correct level of skills.

*WM_RANDOMIZESPELL now casts the correct level of skills.

*Added some needed code for a few skills. Misc type and defense ignoreing ones.

*Updated the following skills: Swing Dance, Rush Windmill, Echo Song, Symphony of Lovers.

*Metallic sound no longer causes Chaos effect, the SP burn of the skill is implemented.

*Fixed a possible crash when clones use Asura Strike.

*Change Material no longer need exact amount of items, it tries to do the change as many times as possible.

+++ sql-files/logs.sql
@@ -60,7 +60,7 @@ CREATE TABLE `mvplog` (
   `kill_char_id` int(11) NOT NULL default '0',
   `monster_id` smallint(6) NOT NULL default '0',
   `prize` smallint(5) unsigned NOT NULL default '0',
-  `mvpexp` mediumint(9) NOT NULL default '0',
+  `mvpexp` bigint(20) unsigned NOT NULL default '0',
   `map` varchar(11) NOT NULL default '',
   PRIMARY KEY  (`mvp_id`)
 ) ENGINE=MyISAM AUTO_INCREMENT=1 ;
